requirements.txt
python-telegram-bot[job-queue]==20.8
zarinpal-python-sdk>=0.2
# Crypto payment dependencies
tronpy>=0.6.1
coincurve>=21.0.0
base58>=2.1.1
eth-abi>=5.2.0
pycryptodome>=3.23.0
pytz==2023.3
python-dotenv==1.0.0
jdatetime>=2.0
qrcode[pil]==7.4.2  # For QR code generation with image support
Pillow==10.3.0 # Image processing library, often a dependency for qrcode image output
requests==2.31.0
openai>=1.12.0
google-generativeai>=0.3.0
python-docx>=1.0.0
faiss-cpu==1.11.0
numpy==2.0.2
tiktoken==0.9.0
nltk==3.9.1
pandas>=2.2.2
XlsxWriter>=3.2.0
openpyxl>=3.1.2
pytest>=8.2.0
